// src/app/features/clientes/models/persona-natural.model.ts
export interface PersonaNatural {
    id: number;
    empresa: string;
    tipoDeIdentificacion: number;
    numeroDeIdentificacion: string;
    primerNombre: string;
    segundoNombre: string;
    primerApellido: string;
    segundoApellido: string;
    nombreCorto: string
    tipoDeTelefono: string;
    numeroDeTelefono: string;
    tipoDeCorreoElectronico: string;
    correoElectronico: string;
}